+++ src/wai/common/json/__init__.py
@@ -1,3 +1,0 @@
-"""
-Package for working with JSON/JSONSchema in an object-oriented manner.
-"""