+++ src/wai/common/json/error/_JSONError.py
@@ -1,5 +1,0 @@
-class JSONError(Exception):
-    """
-    Base class for errors occurring from within the JSON package.
-    """
-    pass