+++ src/wai/common/json/error/_JSONPropertyError.py
@@ -1,8 +1,0 @@
-from ._JSONError import JSONError
-
-
-class JSONPropertyError(JSONError):
-    """
-    Base class for all errors involving JSON properties.
-    """
-    pass