+++ src/wai/common/json/error/_JSONSchemaError.py
@@ -1,8 +1,0 @@
-from ._JSONError import JSONError
-
-
-class JSONSchemaError(JSONError):
-    """
-    Base class for all errors involving JSON schema.
-    """
-    pass