+++ src/wai/common/json/error/_JSONSerialisationError.py
@@ -1,9 +1,0 @@
-from ._JSONError import JSONError
-
-
-class JSONSerialisationError(JSONError):
-    """
-    Class for errors occurring during serialisation/deserialisation
-    of JSON.
-    """
-    pass