+++ src/wai/common/json/error/_JSONValidationError.py
@@ -1,8 +1,0 @@
-from ._JSONError import JSONError
-
-
-class JSONValidationError(JSONError):
-    """
-    Base error type for problems validating JSON.
-    """
-    pass