+++ src/wai/common/json/error/_RequiredDisallowed.py
@@ -1,9 +1,0 @@
-from ._JSONPropertyError import JSONPropertyError
-
-
-class RequiredDisallowed(JSONPropertyError):
-    """
-    Error type for when trying to set a property as required,
-    but it is not allowed.
-    """
-    pass