+++ src/wai/common/json/serialise/__init__.py
@@ -1,7 +1,0 @@
-"""
-Package for interfaces supporting JSON serialisation.
-"""
-from ._JSONBiserialisable import JSONBiserialisable
-from ._JSONDeserialisable import JSONDeserialisable
-from ._JSONSerialisable import JSONSerialisable
-from ._JSONValidatedBiserialisable import JSONValidatedBiserialisable