+++ CHANGES.rst
@@ -8,6 +8,7 @@ Changelog
 - Added function to get the kwargs that are supplied to a function (non-default).
 - Added decorator ensure_error_type which wraps exceptions not of a particular
   type in that type.
+- Removed json package, now in its own repository, wai-json.
 
 0.0.22 (2020-02-12)
 -------------------